FText key stablization
FText DisplayString shared ptrs are persisted across property edits. This preserves their entry in the live table and also updates the string that the entry points to. We call GetString() to update the hashed value as persisting the DisplayString means that its lcoation in the table is preserved.
On serialization, a valid key is still found in the live table for the display string, so the key is preserved.

UE-5350 - As a translator, I would like the editor tutorial text to preserve it's history between changes in order to greatly improve the translation workflow.

reviewed by Saul.Aubreu

[CL 2397482 by Thomas Sarkanen in Main branch]