Changed Web Browser Viewport to use local size so that image is zoomed when application scale is changed.
This also ensures that the mouse positions line up with what's rendered. Could have also been done by changing browser zoom level so that it stays sharp but it would have meant converting from a scale of 1 to a zoom level of 0 and making all mouse events use absolute positions.

[CL 2397639 by Matthew Griffin in Main branch]