Reduced accuracy for angular unit conversion unit test
The problem here seemed to be that PI at runtime is only precise to 1e-7, which seemed to be exacerbated by the unit conversion that resulted in a slightly too large error for the unit test. I've reduced the epsilon for this test a little, and improved the unit test reporting output.

[CL 2399697 by Andrew Rodham in Main branch]