Fixed inline editable text boxes not restoring focus when they are closed
UE-7201 - Crash occurs as user is able to leave text fields in editable state

Closing an inline editable text would remove the editable text box from the widget hierarchy, but not specify where the focus should go. This resulted in list/tile views partially receiving focus, allowing the keyboard to interact with them in undesirable ways.

This change forces the inline editable text box to return focus to whatever had focus when it was opened. For list views, this returns focus to the selected item. For blueprints, this returns focus to the selected node.

ReviewedBy Andrew.Rodham

[CL 2400810 by Jamie Dale in Main branch]