Made text blocks better respect their justification method when the visible size is smaller than the desired size
A text block with a visible size smaller than the desired size of the text layout would previously always favor showing the left side of the text layout. This lead to some odd looking results when using center or right justification, as it was possible for the text to be completely hidden or badly aligned.

This change has the text block "auto-scroll" the text layout to try and stick to the most important part of the text layout (the center when using center justification, and the right when using right justification). This produces much better results for long asset names in the Content Browser (which uses center justified text).

ReviewedBy Andrew.Rodham

[CL 2401816 by Jamie Dale in Main branch]