#github PR748, fix for Log Reverse Attenuation Function from korypostma
Taking suggest fix from pull req but keeping the clamp for Max(foo, 0) because as Distance/Falloff becomes greater than 0.99, we'll get negative attenuations.

Note of course, logf(0.0f) results in -1.#INF but this will test as a negative number (and has precedence in the ATTENUATION_Logarithmic case as well for when Distance == 0.0f).

[CL 2409860 by Aaron McLeran in Main branch]