Fixed context menus on editable texts being unusable due to focus loss
UE-7906 - Right clicking in a text field removes focus from it which makes context menu options not function

The change to set bFocusImmediately to true when summoning the context menus for editable texts meant that we received some focus changed events before we'd set the context menu window pointer (as the call to summon the menu hadn't yet returned).

We now track some extra state to say that we're pending a summon and ignore the focus events in this case (as we do when a context menu is full open).

#codereview Matt.Kuhlenschmidt

[CL 2415303 by Jamie Dale in Main branch]