Made the auto-saved directory path hard-coded
UE-8432 - Crash saving an asset in a project with a long path name

The auto-save dir used to be saved in the editor config file, but was always the same and couldn't be moved outside of the "Saved" directory. This was brought up due to the config issues raised by UE-8542, so we've now agreed to just hard-code this path.

[CL 2423900 by Jamie Dale in Main branch]