Fixed a crash caused by an image run being left as the only run in a text layout
UE-7717 - Crash when deleting an image added via an undo in editable rich-text

FTextLayout::RemoveAt was trying to ensure that a line never had zero runs by preventing the removal of the last run on a line, however, this can cause a crash if the last run on a line is not a text-based run.

This change makes FTextLayout::RemoveAt always remove a run when it becomes empty, but then re-add a new default text run if the removal caused the line to have zero runs.

ReviewedBy Justin.Sargent

[CL 2423959 by Jamie Dale in Main branch]