Fixed UCanvas::MeasureStringInternal sometimes overmeasuring strings
UE-7427 - In game: Console font doesn't match to cursor poistion

UFont::GetCharSize was returning the width as the X advance, plus the horizontal offset, however the horizontal offset is only used as an adjuster when working out the center of a character.

UFont::GetCharSize has been changed to set the width as just the X advance, and I've added UFont::GetCharHorizontalOffset to get the horizontal offset as required when using UCanvas::MeasureStringInternal with ELastCharacterIndexFormat::CharacterAtOffset.

[CL 2425558 by Jamie Dale in Main branch]