We no longer limit the window size when reshaping a fullscreen window
UE-9036 - Changes to Resolution are not saved if Fullscreen is toggled on

Windows defaults to preventing you being able to size a window larger than the currently available desktop size. When running in D3D fullscreen mode, this means that it prevents you being able to resize a window larger than the current fullscreen resolution.

You can override this limitation by passing SWP_NOSENDCHANGING when calling SetWindowPos. This prevents the window being notified that it is about to change size (preventing the WM_GETMINMAXINFO message), and instead just notifies the window of its new size.

This CL also makes sure we call SetWindowPos after changing the window styles, as the MSDN stipulates that the DWM may cache this information until notified of a change by SWP_FRAMECHANGED.

ReviewedBy Matt.Kuhlenschmidt

[CL 2438301 by Jamie Dale in Main branch]