Renamed Actor Locking editor functionality to Actor Pilot
Actor pilots are now initiated from the actor's context menu, rather than an unrelated menu on the viewport. Similarly, pilots can be stopped via the context menu for the actor, or from the viewport toolbar that is visible when a pilot is active.

This addresses UE-1091 - Allow actors to be piloted with the camera in UE4.

[CL 2446881 by Andrew Rodham in Main branch]