Added support for "gathering" project files via the editor
This is needed for UBT to update its makefiles when we add new code via the editor.

Updating UBT makefiles is more typically handled by re-generating project files, however on Windows we use DTE to inject new files directly into Visual Studio in an attempt to avoid re-generating project files (as it produces modal popups that block operations), so we need to perform a gather instead so that UBT knows things have changed.

ReviewedBy Mike.Fricker

[CL 2446903 by Jamie Dale in Main branch]