Added the ability to derive a Blueprint from a C++ class via the Content Browser
UE-9293 - Should offer 'Create BP' when right clicking on C++ class in Browser

There is now a "Create Blueprint class based on X" option alongside the "Create C++ class derived from X" option when viewing the context menu for a C++ Class in the Content Browser. This allows you to create a new Blueprint asset from C++ classes that are marked as blueprintable.

[CL 2447345 by Jamie Dale in Main branch]