Fixed unloaded Blueprint components having the wrong icon
UE-9331 - Missing icons for BP component classes when restarting editor

Moved the code to work out the correct class icon for a Blueprint or asset data into FClassIconFinder so that it can be shared between the asset thumbnails and the components context menu.

Added an extra IconClass member to FComponentClassComboEntry so that unloaded Blueprints can set this appropriately despite having a null ComponentClass.

[CL 2451813 by Jamie Dale in Main branch]