Changed editor feedback contexts to always delete the progress window if it exists.
There was a potential vulnerability where a slow task widget would not get cleaned up resulting in it outliving the pointer to the scope stack. The context no longer holds a strong reference to the slow task widget in a bid to prevent it from outliving the data to which it is tied.

[CL 2455914 by Andrew Rodham in Main branch]