UE-8676 Crash when changing a Blueprint's vector variable to a transform variable UE-9089 Better handling for UDS rename
Problem: When a property was of type on native struct (that is serialized binary) and the type was change to user defined struct (that is serialized using tags): the new structure tries to serialize tag from binary data of the old struct.
Solution: Use persistent guid to identify user defined structures (even after they are renamed).

#codereview Nick.Whiting, Robert.Manuszewski

[CL 2464813 by Maciej Mroz in Main branch]