Fixed text export of UModel objects, so that Brush objects are forward declared before being defined. #jira UE-1264 - LICENSEE: UDN: EDITOR: Multiple Warning: Ambiguous search messages clogging output log
UModel objects were not being forward declared, meaning that the reference in the BrushComponent was ambiguous (and also referring to the wrong object).  By luck, this was being fixed up at the end of ULevelFactory::FactoryCreateText for this particular case, but other exported UModel* properties would likely fail.  This change correctly forward declares any UModel objects in the export text and allows them to be redefined during import if they are uninitialized.

#reviewedby Matt.Kuhlenschmidt

[CL 2472016 by Richard TalbotWatkin in Main branch]