Core: Added support for variadic template parameters in TypeContainer for TFunction and Delegate registrations.
The syntax is not ideal yet as explicit specification of template parameters is still required. Automatic deduction of these will require much more meta-programming, but is probably doable. Also, the delegate registration with variadic parameters currently requires the delegate type to be specified as well. Other than that, high fives all around!

#CodeReview: justin.sargent

[CL 2474270 by Max Preussner in Main branch]