Fix deleted behavior tree blueprint nodes showing up in context menu
By the time assets are deleted, ResolveName() will not return the object name as the underlying package is no longer there, so it will not get deleted form the class tree. Performing direct string manupulation on the object path avoids this.

UE-8201 - Task node blueprints remain active in the context menu after they are deleted.

[CL 2482961 by Thomas Sarkanen in Main branch]