Tests ran using -automationtests= will now run after the asset registry has finished loading.  This does mean that on some larger projects the tests may take longer before starting.  This is better than not testing them at all which was the case previously.
UnrealEngine.cpp - removed the automationtests commandline arguments from being added to the deferredcommands array.
AutomationWorkerModules - When initailized the commandlines are added to the new deferredautomationcommands array.  Wrote a function to exectue the commands.  Bound the function to the asset registry delegate that gets called when the the registry is done loading.

[CL 2483033 by Justin Hair in Main branch]