Implemented Page Up/Down support for multi-line editable text
UE-1496 - Slate: Implement PageUp PageDown support for multiline editable text

This will move the multi-line text by a "page" (based on the number of visible lines), and then scroll the document to try and keep the cursor at the same relative location in the view.

This also fixes some issues with scrolling when dealing with a scaled multi-line editable text, as the scroll offset is now always considered to be unscaled until it is given to the text layout. This prevents issues where we could over or under scroll when using a scale other than 1.

#codereview Justin.Sargent

[CL 2485106 by Jamie Dale in Main branch]