Fixed cached source file paths not being updated when they had changed.
Previously the asset registry skipped the code to update cached tags is the keys were the same, but we need this code to be updated if the values of some tags are changed. The new code should be negligibly slower in the case where tags haven't changed, and potentially faster when they have.

Previously submitted in 2483050, but the change went awry.

[CL 2485323 by Andrew Rodham in Main branch]