Auto Reimport - file hashes are now harvested on a background thread once the initial directory scan has completed.
This has the benefit of startup times being hugely improved in the worse case when no cache was present.

The downside is that rename/move detection on live changes may not work if the hashes have not been gathered yet, but this is only an issue if there was no hash, and the changes happen soon after the editor starts up. Rename/Move changes that apply to files with that haven't had a hash gathered will just be reported as removes and adds.

[CL 2485355 by Andrew Rodham in Main branch]