Data table UI improvements
UETOOL-291 - Improve the data table UI

The column sizes in the data table can now be resized as desired, and also default to being sized to match their content rather than as a simple proportion of their parent size. Any resized columns are remembered between editing sessions.

The data table view now scrollable horizontally as well as vertically to account for the fact that the data may now be wider than the current view.

[CL 2495732 by Jamie Dale in Main branch]