Added the ability to re-order rows in the data table editor
UETOOL-295 - Can't re-order rows in a data table

Added an option to move a row up or down, or to the top or bottom of the data table.

Updated the row editor UI to be more consistent with other editors, and re-arranged some of the controls on the toolbar, adding tooltips where appropriate.

Fixed the highlighted row in the data table not matching the selected row editor row when a data table was first opened.

[CL 2497325 by Jamie Dale in Main branch]