Math expression nodes now support variable renames without breaking and local variables can be used with them.
Behind the scenes, variables are converted to their Guid and are converted back to their name for display. The token parser was updated to handle these Guid's which are inside brackets { }.

Added some helper functions for looking up a Member Variable's Guid by name and a Member Variable's name by Guid.

#jira UE-9915 - Renaming Blueprint variables breaks MathExpression node on restart
#jira UE-12363 - Math expression node doesn't work with local variables.

[CL 2500113 by Michael Schoell in Main branch]