PR #868: Add "Snap Object To View" to Editor (Contributed by tommybear)
This allows you to align your view in the editor, select an actor in the
world outliner or in scene, right-click and select "Snap Actor to
Camera".  Actor will then set its location and rotation to match the
view.  This is super handy for quickly finding an awesome camera view in
the editor and then making your CameraActors match that view.

[CL 2501200 by Matt Kuhlenschmidt in Main branch]