Added a way for STableViewBase (and its derived types) to notify you when they are scrolled.
This allows you, with the combination of SetScrollOffset, to sync two or more list views to have the same scroll offset.

STableViewBase no-longer associates its header with an external scrollbar. This prevents the header from trying to compensate for something that doesn't exist, and causing the header and the rows to become out of line.

ReviewedBy Nick.Atamas

[CL 2508476 by Jamie Dale in Main branch]