Converted the data table editor to use SListView
UETOOL-281 - Improve data table workflows

This is needed as the changes that had been made to SSplitter have been moved to SHeaderRow instead, as they were too specific to go into the generic splitter.

This means that the data table editor now has two list views (one for the row names, one for everything else) which are kept in sync. We have two so that the row names can be anchored to the left of the screen when scrolling the other columns.

[CL 2508489 by Jamie Dale in Main branch]