Improved the error message when adding a new class fails due to a compile error
The fact that the error message said "failed to add class" despite the class itself being added was causing confusion.

In the case where adding a new class fails to hot-reload (probably due a compiler error) we now state that the class itself has been added, but that you'll need to recompile the module before it will appear in the editor, and offer an easy way to open the output log to see more detailed compiler output. We also automatically close the class wizard so that you're not left looking at an error message saying that "the class name is already in use".

[CL 2515833 by Jamie Dale in Main branch]