Changed editor project unit settings to be more flexible
Users can now choose a set of relevant units to display applicable properties in.
Setting a single unit for a given type will result in only that unit ever being used. When multiple exist, the most appropriate unit will be shown when a property editor is opened, and will remain in this unit until something else is selected.

[CL 2519064 by Andrew Rodham in Main branch]