Fixed editor per-project user settings getting eradicated on startup
The editor per-project settings in some cases were not being loaded correctly when bGenerateDestIni was false, which caused the file to be altogether emptied. I've changed this code to load the config directly, rather than rely on the nuanced behavior of the FConfigCacheIni helper functions. Also change the code to back up the old settings, rather than destroy them.

[CL 2529955 by Andrew Rodham in Main branch]