Fixed the game mode class always showing as "None" in the project settings
UE-15153 - Unable to change Default Game Mode through Project Settings window

CL# 2483623 (to fix UE-11234) made GetCurrentGameModeClass use GetValue to get the value as an object, rather than get the value as a string and then use FindObject to locate the class, however the property handle used by FStringClassReference doesn't implement a GetValue function that returns an object, so the function was always failing and returning null.

I've restored the code back to its original state, and added a guard against calling FindObject while GIsSavingPackage is true. I also added a cached version of the game mode class so that we can try and return something valid in the case where GetCurrentGameModeClass is called while GIsSavingPackage is true.

[CL 2545444 by Jamie Dale in Main branch]