Can set Function and Macro graph metadata for CompactNodeTitle and Keywords.
Changed the StandardMacros Increment, Decrement, and Negate functions to use this metadata and removed the code forcing them to appear as compact nodes from UK2Node_MacroInstance.

Compact Node Title will force nodes to display as Compact Nodes if set. This information is also appended to the Keywords, no need to set in the Keywords as well.

Keywords makes nodes easier to find.

#jira UE-14499 - Setting Function/Macro graph Keyword and Compact Title metadata

[CL 2545723 by Michael Schoell in Main branch]