Behavior Tree debugger improvements
UETOOL-298 - The "step into"/out/over buttons in the Behavior Tree enable and disable all the time

The "Back: Over" and "Forward: Over" step options are now able to perform a step that means they move back up to a parent tree. Previously this was being disallowed as the they were incorrectly treating that as entering a sub-tree.

If you pause your PIE session, the behavior tree debugger will now make sure and open the correct graph. Previously you could pause while within a sub-tree and the view would be left on the parent graph.

[CL 2547021 by Jamie Dale in Main branch]