Changed the USceneComponent::PhysicsVolume property to a TWeakObjectPtr. #jira UE-15435 - Attempting to open a map in content browser that is currently loaded in world composition browser crashes editor
Previously, the PhysicsVolume pointer was creating a dependency between sublevels and the persistent level as its default value was the default physics volume of the persistent level, no matter which level the component was in.  This was causing GC fails when loading a sublevel which was already loaded.  Changing to a TWeakObjectPtr breaks this dependency.

#reviewedby Zak.Middleton

[CL 2549415 by Richard TalbotWatkin in Main branch]