Removed the animation from SEditableText so that Slate can sleep
UE-11632 - Make SEditableText sleep-friendly.

SEditableText used to have spring animations for moving the caret and changing the selection, however these animations required the editable text to keep Slate active while they had focus in order to smoothly update the animation.

These animations have been removed, and the only ticking the editable text now needs to do is to blink the cursor (which is a much coarser animation). This also brings the look-and-feel of SEditableText and SMultiLineEditableText closer together, as the latter has never used these spring animations.

[CL 2555414 by Jamie Dale in Main branch]