Fixed "Delete" on an empty line in ML text being handled like a "Backspace"
The ML text logic for using delete on an empty line was trying to ensure that the first line wasn't removed, and would then move the cursor back to the end of the previous line. This is incorrect, as that is the behavior of the backspace key - delete should make sure that the last line can't be removed, and shouldn't move the cursor at all.

#codereview Justin.Sargent

[CL 2556834 by Jamie Dale in Main branch]