Fix for CRASH: EDITOR: Crash occurs when connecting MF_DepthEnv to the M_Forest Material Attributes
Several recent additions to the material properties have made the mapping from IO index to EMaterialPropertiy very confusing and fragile.
I've pulled the mapping to a single place in the code and it should now be robust to new additions.
Changing the order of pins will still require fix up of old FExpressionInputs but this is true of all expressions. No way around that.
Though I've now placed some static asserts and comments around the code so that the need for this and how to easily avoid it at all is clear and unmissable.

[CL 2565623 by Simon Tovey in Main branch]