UnrealAudio Soundfile encoding and decoding
- Implemented libsndfile soundfile class for encoding and decoding audio files (.wav, .aif, .ogg, .flac)
- Implemented import/export process tasks as async background tasks
- Added testing code to test single-file importing and exporting
- Added testing code to test batch-file importing and exporting
- Add dlls for Win32/Win64/Mac. Tested on Win64 and Mac.

To test file import/export processing, use the AudioTestCommandlet with the following:
-command UnrealEd.AudioTestCommandlet source export [SingleFile/Directory]

Where [SingleFile/Directory] is the full path to either a single audio file or a directory to a folder containing audio assets supported. All files are imported into an internal binary ogg format (I will make it variable import format to test), then immediately exported back to the original format. If a file fails importing due to a file format issue (or some other issue), that issue will be reported in the log and the batch processing will proceed.

The exported file will be exported into the original format and put in a Exported sub-directory of the directory of the given single file or the import directory with "_exported" appended to the name.

[CL 2565979 by Aaron McLeran in Main branch]