Added quick asset management for collections in the editor
UETOOL-332 - Collections 2.0
UETOOL-370 - Ability to assign many collections to a single asset very quickly

The collections view now provides check boxes next to each collection. These can be used to quickly add or remove a large selection of assets to or from a collection.

This feature is also available in the Content Browser via the asset context menu. The "Collections" section now contains a sub-menu which lists the available collections, and allows you to quickly add or remove the current selection to or from a collection.

[CL 2570749 by Jamie Dale in Main branch]