Can right click on any node to search for references of it. Does initial search only in current Blueprint.
Localization fixes for Find-in-Blueprints so it will again search the current and native language.

CallFunction and CustomEvents now store their non-friendly "native" names. These will not actually appear in search results, but will be compared while searching to determine if the node matches the results.

Gathering Blueprint search data will now force friendly names to true while it gathers so gathered data does not differ based on the setting.

#jira UE-15901 - Right click a custom event to get a list of all blueprints that are calling the custom event

[CL 2572814 by Michael Schoell in Main branch]