Expression parser improvements
Refactored the guts of FExpressionNode to better afford holding moveable types on the stack. Large types are now also stored in the inline bytes for simplicity's sake, but allocated on the heap, and managed with a TUniquePtr.
Added the concept of an evaluation context that can be passed in to the Evaluate function. Jump tables are now templated on a given context type, which is wrapped up with a supplied context for evaluation purposes into a TOperatorEvaluationEnvironment (so that the evaluation logic need not be concerned with the context itself)

[CL 2576821 by Andrew Rodham in Main branch]