Added advanced search query syntax to the Content Browser
UETOOL-332 - Collections 2.0
UETOOL-366 - Ability to search and filter assets that are only in specific collections

This allows you to perform complex queries against any of the meta-data from the asset registry (using standard comparison operators), as well as being able to test other standard keys, such as the collections that contain a given asset, its name, path, or class.

The syntax supports sub-expression grouping, and also partial (unquoted), exact (quoted) starts with (s...), or ends with (...s) string matching, as well as standard logic operators (OR, AND, NOT).

ReviewedBy Andrew.Rodham

[CL 2578624 by Jamie Dale in Main branch]