Improved handling of quotes within an unquoted string
If the text filter expression parser comes across a quote within a quoted string, it will now switch into "quoted string" parsing mode, and consume all characters (even special characters) until it finds a corresponding closing quote. The resultant string will also include the quote characters (unlike a normal quoted string).

This allows our asset reference strings to be parsed correctly, eg) Type'/Path/To/Asset.Asset'

This also improves the error message when a quoted string is unbound. It now reports the character it was looking for, as well as the line and character index of the opening quote.

ReviewedBy Andrew.Rodham

[CL 2581449 by Jamie Dale in Main branch]